Revert "Fixed bug preventing creation of new news"
This reverts commit dfbd781d59d984b1a32d13aa7f1d99d6497958f8.

+++ Ålleinfo Admin/NewsItem.cs
@@ -134,7 +134,7 @@ namespace Ålleinfo_Admin
             {
                 try
                 {
-                    Webber.removeNews((int)data.id);
+                    Webber.removeNews(data.id);
 
                     AdminForm.adminForm.removeNewsDisplay(this);
                 }
